+++ bigchaindb/util.py
@@ -1,4 +1,7 @@
 import multiprocessing as mp
+import json
+import time
+from datetime import datetime
 
 
 class ProcessGroup(object):
@@ -21,4 +24,36 @@ class ProcessGroup(object):
                               kwargs=self.kwargs, daemon=self.daemon)
             proc.start()
             self.processes.append(proc)
+
+
+def serialize(data):
+    """Function used to serialize a dict into a JSON formatted string.
+
+    This function enforces rules like the separator and order of keys. This ensures that all dicts
+    are serialized in the same way.
+
+    This is specially important for hashing data. We need to make sure that everyone serializes their data
+    in the same way so that we do not have hash mismatches for the same structure due to serialization
+    differences.
+
+    Args:
+        data (dict): dict to serialize
+
+    Returns:
+        str: JSON formatted string
+
+    """
+    return json.dumps(data, skipkeys=False, ensure_ascii=False,
+                      separators=(',', ':'), sort_keys=True)
+
+
+def timestamp():
+    """Function to calculate a UTC timestamp with microsecond precision.
+
+    Returns:
+        str: UTC timestamp.
+
+    """
+    dt = datetime.utcnow()
+    return "{0:.6f}".format(time.mktime(dt.timetuple()) + dt.microsecond / 1e6)
