+++ tests/utils/test_config_utils.py
@@ -3,6 +3,7 @@ import copy
 import pytest
 
 import bigchaindb
+from bigchaindb import exceptions
 
 
 ORIGINAL_CONFIG = copy.deepcopy(bigchaindb.config)
@@ -34,5 +35,5 @@ def test_bigchain_instance_raises_when_not_configured(monkeypatch):
     # from existing configurations
     monkeypatch.setattr(config_utils, 'autoconfigure', lambda: 0)
 
-    with pytest.raises(bigchaindb.core.KeypairNotFoundException):
+    with pytest.raises(exceptions.KeypairNotFoundException):
         bigchaindb.Bigchain()
